+++ tests/HelpersTest.php
@@ -13,11 +13,11 @@ declare(strict_types=1);
 
 namespace GrahamCampbell\Tests\TestBenchCore;
 
-use Exception;
 use GrahamCampbell\TestBenchCore\HelperTrait;
 use InvalidArgumentException;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
+use TypeError;
 
 class HelpersTest extends TestCase
 {
@@ -25,83 +25,63 @@ class HelpersTest extends TestCase
 
     public function testInArray()
     {
-        $this->assertInArray('foo', ['foo']);
+        self::assertInArray('foo', ['foo']);
     }
 
     public function testNotInArray()
     {
-        $this->expectException(ExpectationFailedException::class);
-        $this->assertInArray('foo', ['bar']);
+        self::expectException(ExpectationFailedException::class);
+        self::assertInArray('foo', ['bar']);
     }
 
     public function testMethodDoesExist()
     {
-        $this->assertMethodExists('getBar', FooStub::class);
+        self::assertMethodExists('getBar', FooStub::class);
     }
 
     public function testMethodDoesNotExist()
     {
-        $this->expectException(ExpectationFailedException::class);
-        $this->assertMethodExists('getFoo', FooStub::class);
+        self::expectException(ExpectationFailedException::class);
+        self::assertMethodExists('getFoo', FooStub::class);
     }
 
     public function testInJson()
     {
-        $this->assertInJson('{"foo":"bar"}', ['foo' => 'bar']);
-        $this->assertInJson('{ "foo": "bar", "bar": "baz" }', ['foo' => 'bar']);
-        $this->assertInJson('{ "foo": "bar", "bar": "baz" }', ['bar' => 'baz']);
+        self::assertInJson('{"foo":"bar"}', ['foo' => 'bar']);
+        self::assertInJson('{ "foo": "bar", "bar": "baz" }', ['foo' => 'bar']);
+        self::assertInJson('{ "foo": "bar", "bar": "baz" }', ['bar' => 'baz']);
     }
 
     public function testNotInJsonOne()
     {
-        $this->expectException(ExpectationFailedException::class);
-        $this->assertInJson('{"foo":"baz"}', ['foo' => 'bar']);
+        self::expectException(ExpectationFailedException::class);
+        self::assertInJson('{"foo":"baz"}', ['foo' => 'bar']);
     }
 
     public function testNotInJsonTwo()
     {
-        $this->expectException(ExpectationFailedException::class);
-        $this->assertInJson('{ "foo": "bar", "bar": "baz" }', ['foo' => 'baz']);
+        self::expectException(ExpectationFailedException::class);
+        self::assertInJson('{ "foo": "bar", "bar": "baz" }', ['foo' => 'baz']);
     }
 
     public function testBadJsonInNotInJson()
     {
-        $this->expectException(InvalidArgumentException::class);
-        $this->assertInJson('foobar', ['foo' => 'bar']);
+        self::expectException(InvalidArgumentException::class);
+        self::assertInJson('foobar', ['foo' => 'bar']);
     }
 
     public function testAssertArraySubsetBadArg1()
     {
-        try {
-            $this->assertArraySubset(123, []);
-        } catch (Exception $e) {
-            $this->assertInArray($e->getMessage(), [
-                'Argument #1 of GrahamCampbell\Tests\TestBenchCore\HelpersTest::assertArraySubset() must be an array or ArrayAccess',
-                'Argument #1 (No Value) of GrahamCampbell\Tests\TestBenchCore\HelpersTest::assertArraySubset() must be a array or ArrayAccess',
-                'Argument #1 (No Value) of PHPUnit\Framework\Assert::assertArraySubset() must be a array or ArrayAccess',
-            ]);
+        $this->expectException(TypeError::class);
 
-            return;
-        }
-
-        $this->assertTrue(false);
+        self::assertArraySubset(123, []);
     }
 
     public function testAssertArraySubsetBadArg2()
     {
-        try {
-            $this->assertArraySubset([], 123);
-        } catch (Exception $e) {
-            $this->assertInArray($e->getMessage(), [
-                'Argument #2 of GrahamCampbell\Tests\TestBenchCore\HelpersTest::assertArraySubset() must be an array or ArrayAccess',
-                'Argument #2 (No Value) of GrahamCampbell\Tests\TestBenchCore\HelpersTest::assertArraySubset() must be a array or ArrayAccess',
-                'Argument #2 (No Value) of PHPUnit\Framework\Assert::assertArraySubset() must be a array or ArrayAccess',
-            ]);
+        $this->expectException(TypeError::class);
 
-            return;
-        }
-
-        $this->assertTrue(false);
+        self::assertArraySubset([], 123);
     }
 }